requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.7
dal-admin-filters==1.1.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.2
Django==4.2.10
django-allauth==0.60.1
django-autocomplete-light==3.12.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-import-export==3.3.6
django-jazzmin==2.6.0
django-multiupload==0.6.1
django-redis==5.4.0
django-timezone-field==6.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
prompt_toolkit==3.0.44
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
